gpu: Improve DVFS utilization accuracy
Get the current time at the same time the utilization registers are
accessed, and disable interrupts on RT while it happens so that nothing can
interrupt the calculation and hurt the accuracy of the numbers.

This ensures that inaccurate DVFS utilization numbers aren't fed into the
governor, which should therefore improve the frequency selections made by
the governor.

Change-Id: I92d23df816750ea0b5adb7e4b69a2873a4779f14
Signed-off-by: electimon <electimon@gmail.com>